Progress on final
Added main.py, logo, inventory.xlsx, and requirements.txt. Just posted kind of my take on it. Still requires work to get the excel button to work correctly, I would also much rather write to a database so the program adds to it but wasn't able to get that all figured out. Everything should be well documented in the comments.